cs4300/homework2/movie_theater_booking/requirements.txt
asgiref==3.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.7
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.11.0
